backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
python-dotenv
PyMuPDF
openai
pinecone-client
pytesseract
pillow
python-multipart
